requirements.txt
# Python dependencies for Lumina Photo Enhancer
# Run: pip install -r requirements.txt

# Core Web Framework
fastapi
uvicorn[standard]
pydantic

# HTTP client
httpx

# Google Gemini AI
google-genai

# Async HTTP (for GenAI SDK)
aiohttp

# Settings
python-dotenv

# File handling
aiofiles
python-multipart

# WebSocket support (included in uvicorn[standard])
websockets
